requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0
